Final project first push
Ran XFLR5 with TTL at 40m/s and STP.  Varied AOT from -5 to 5 and
exported .txt files with Cp and Xcp that may have the data used in the
paper.  This was the analysis configuration saved on the file.  Also
inlcuded is a xcel sheet showing the wing configurations and notes of
what was held constant in each.